Beginnings of an evaluation module.
Includes:
 - A refactor of solution.py to work with multiple filters.
 - Consolidation of tools
 - Quantitative evaluation as well as visual evaluation in compare.py.
