+++ gnss_analysis/filters/swiftnav_filter.py
@@ -12,7 +12,7 @@ import pandas as pd
 from swiftnav import dgnss_management
 
 from gnss_analysis import dgnss, ephemeris
-from gnss_analysis.filters.common import TimeMatchingDGNSSFilter
+from gnss_analysis.filters.dgnss_filter import TimeMatchingDGNSSFilter
 
 
 class SwiftNavDGNSSFilter(TimeMatchingDGNSSFilter):
@@ -57,6 +57,18 @@ class SwiftNavDGNSSFilter(TimeMatchingDGNSSFilter):
     # update the ambiguities
     dgnss_management.dgnss_update_ambiguity_state_(self.amb_state)
 
+    flag, n_used, baseline = dgnss_management.dgnss_baseline_(sdiff_t,
+                                              self.base_pos,
+                                              self.amb_state,
+                                              disable_raim=self.disable_raim)
+    # check the flag to see if baseline computations succeeded.  If not
+    # we return None instead of a possibly bogus baseline.
+    if flag < 0:
+      logging.warn("Baseline computation failed with %d" % flag)
+    else:
+      self.baseline = pd.Series(baseline, index=['x', 'y', 'z'])
+
+
   def get_baseline(self, obs_set):
     """
     Uses the current filter state to estimate the baseline for
@@ -65,23 +77,8 @@ class SwiftNavDGNSSFilter(TimeMatchingDGNSSFilter):
     Base observations are propagated to the rover time of arrival
     (if required) before being used to compute the baseline.
     """
-    # Add the satellite state information (position etc ..)
-    # to the rover observations.
-    obs_set['base'] = ephemeris.add_satellite_state(obs_set['base'],
-                                                    obs_set['ephemeris'])
-    # creates a DataFrame of single differences
-    sdiffs = self.get_single_diffs(obs_set['rover'], obs_set['base'],
-                                   propagate_base=True)
-    # converts the DataFrame to c objects
-    sdiff_t = list(dgnss.create_single_difference_objects(sdiffs))
-    # use the current filter state to estimate the baseline
-    flag, n_used, baseline = dgnss_management.dgnss_baseline_(sdiff_t,
-                                                  self.base_pos,
-                                                  self.amb_state,
-                                                  disable_raim=self.disable_raim)
-    # check the flag to see if baseline computations succeeded.  If not
-    # we return None instead of a possibly bogus baseline.
-    if flag < 0:
-      logging.warn("Baseline computation failed with %d" % flag)
+    if not self.initialized:
       return None
-    return pd.Series(baseline, index=['x', 'y', 'z'])
+    else:
+      # For now we make just return the current filter state.
+      return self.baseline
